new vm mod
